Move CMS Controller to WizaplaceFrontBundle
Also refactor some code into a service
