+++ src/WizaplaceFrontBundle/Resources/views/cms/page.html.twig
@@ -1,0 +1,5 @@
+<h1>Override me!</h1>
+<p>
+    I'm <i>{{ _self }}</i><br />
+    Context: {{ dump(_context) }}
+</p>
